8. Filing of leave -  when a user selects DATE of leave BEFORE identifying the LEAVE TYPE, it overrules the "3days before filing" set up.
5. User password - on May 1, users should be prompted by the
system to change their password. Otherwise, we all have the same
passwords, and that cannot be. Currently, upon log-in, there's no
way for us to change password.

4. Filed OT - details of approved OT should be on the dashboard
of filer and approver. In the case of Fulfillment Center, Elaine
requests to have the ability to approve/decline all FC filed OTs.
Approver should be Elaine AND supervisor.

3. Approved leaves do not appear on the calendar - last
meeting,
you said that it should appear, but it does not.